requirements.txt
redis==4.5.1
PyYAML==6.0
boto3==1.26.99
slack-sdk==3.20.0